Move FileStuff and TimeStuff to cbot.util; Rename UI to Talker (in previous commit)
