+++ src/main/java/cbot/util/FileStuff.java
@@ -1,4 +1,4 @@
-package cbot.io;
+package cbot.util;
 
 import java.io.File;
 import java.io.FileNotFoundException;
